requirements.txt
pyspark>=3.5.0
pandas>=2.0.0
pyyaml>=6.0
pyarrow>=15.0.0
